Refactor fs and cleanup
