+++ src/extensions/magic_mount.rs
@@ -1,149 +1,43 @@
-use tonic::transport::{Channel, ClientTlsConfig};
-use crate::application::{CoLinkClient, generate_request};
-use crate::colink_proto::*;
+mod chunk;
 
-const CHUNK_SIZE: usize = 1 * 1024 * 1024; // use 1MB chunks
+use crate::colink_proto::*;
 
 type Error = Box<dyn std::error::Error + Send + Sync + 'static>;
 
 impl crate::application::CoLink {
-    pub async fn _create_entry_bulk(
-        &self,
-        key_name: &str,
-        payload: &[u8],
-    ) -> Result<String, Error> {
-        let mut client = self._grpc_connect(&self.core_addr).await?;
-        let metadata_key = format!("{}::metadata", key_name);
-        let metadata_key_copy = metadata_key.clone();
-        let result = metadata_key.clone();
-        let num_chunks = payload.len() / CHUNK_SIZE + 1;
-        let create_request = generate_request(
-            &self.jwt,
-            StorageEntry {
-                key_name: metadata_key,
-                payload: (num_chunks.to_string() + "-locked").into_bytes(), // simply store the number of chunks in the metadata
-                ..Default::default()
-            },
-        );
-        client.create_entry(create_request).await?;
-
-        let mut offset = 0;
-        let mut chunk_id = 0;
-        while offset < payload.len() {
-            let chunk_size = if offset + CHUNK_SIZE as usize > payload.len() {
-                payload.len() - offset
-            } else {
-                CHUNK_SIZE as usize
-            };
-            let chunk = payload[offset..offset + chunk_size].to_vec();
-            let request = generate_request(
-                &self.jwt,
-                StorageEntry {
-                    key_name: format!("{}::{}", key_name, chunk_id),
-                    payload: chunk,
-                    ..Default::default()
-                },
-            );
-            client.create_entry(request).await?;
-            offset += chunk_size;
-            chunk_id += 1;
+    pub async fn _mm_create_entry(&self, key_name: &str, payload: &[u8]) -> Result<String, Error> {
+        let token = key_name.split("$").last().unwrap();
+        if token.eq("chunk") {
+            let response = self._create_entry_chunk(key_name, payload).await?;
+            return Ok(response);
         }
-
-        let update_request = generate_request(
-            &self.jwt,
-            StorageEntry {
-                key_name: metadata_key_copy,
-                payload: (num_chunks.to_string() + "-unlocked").into_bytes(),
-                ..Default::default()
-            },
-        );
-        client.update_entry(update_request).await?;
-        Ok(result)
+        return Err("invalid storage option".into());
     }
 
-    pub async fn _read_entry_bulk(&self, key_name: &str) -> Result<Vec<u8>, Error> {
-        let mut client = self._grpc_connect(&self.core_addr).await?;
-        let metadata_key = format!("{}::metadata", key_name);
-        let request = generate_request(&self.jwt, StorageEntries {
-            entries: vec![StorageEntry {
-                key_path: metadata_key,
-                ..Default::default()
-            }],
-        });
-        let response = client.read_entries(request).await?;
-        let payload_string = String::from_utf8(response.get_ref().entries[0].payload.clone())?;
-        let num_chunks = payload_string.split("-").next().unwrap().parse::<usize>()?;
-        let locked = payload_string.split("-").last().unwrap().eq("locked");
-        if locked {
-            return Err("entry is locked".into());
+    pub async fn _mm_read_entry(&self, key_name: &str) -> Result<Vec<u8>, Error> {
+        let token = key_name.split("$").last().unwrap();
+        if token.eq("chunk") {
+            let response = self._read_entry_chunk(key_name).await?;
+            return Ok(response);
         }
-
-        let mut payload = Vec::new();
-        for chunk_id in 0..num_chunks {
-            let request = generate_request(&self.jwt, StorageEntries {
-                entries: vec![StorageEntry {
-                    key_path: format!("{}::{}", key_name, chunk_id),
-                    ..Default::default()
-                }],
-            });
-            let response = client.read_entries(request).await?;
-            payload.extend(response.get_ref().entries[0].payload.clone());
-        }
-        Ok(payload)
+        return Err("invalid storage option".into());
     }
 
-    pub async fn _update_entry_bulk(
-        &self,
-        key_name: &str,
-        payload: &[u8],
-    ) -> Result<String, Error> {
-        let mut client = self._grpc_connect(&self.core_addr).await?;
-        let metadata_key = format!("{}::metadata", key_name);
-        let metadata_key_copy = metadata_key.clone();
-        let result = metadata_key.clone();
-        let num_chunks = payload.len() / CHUNK_SIZE + 1;
-        let request = generate_request(
-            &self.jwt,
-            StorageEntry {
-                key_name: metadata_key,
-                // store number of chunks and lock in metadata
-                payload: (num_chunks.to_string() + "-locked").into_bytes(),
-                ..Default::default()
-            },
-        );
-        client.update_entry(request).await?;
-
-        let mut offset = 0;
-        let mut chunk_id = 0;
-        while offset < payload.len() {
-            let chunk_size = if offset + CHUNK_SIZE as usize > payload.len() {
-                payload.len() - offset
-            } else {
-                CHUNK_SIZE as usize
-            };
-            let chunk = payload[offset..offset + chunk_size].to_vec();
-            let request = generate_request(
-                &self.jwt,
-                StorageEntry {
-                    key_name: format!("{}::{}", key_name, chunk_id),
-                    payload: chunk,
-                    ..Default::default()
-                },
-            );
-            client.update_entry(request).await?;
-            offset += chunk_size;
-            chunk_id += 1;
+    pub async fn _mm_update_entry(&self, key_name: &str, payload: &[u8]) -> Result<(), Error> {
+        let token = key_name.split("$").last().unwrap();
+        if token.eq("chunk") {
+            let response = self._update_entry_chunk(key_name, payload).await?;
+            return Ok(response);
         }
+        return Err("invalid storage option".into());
+    }
 
-        let update_request = generate_request(
-            &self.jwt,
-            StorageEntry {
-                key_name: metadata_key_copy,
-                payload: (num_chunks.to_string() + "-unlocked").into_bytes(),
-                ..Default::default()
-            },
-        );
-        client.update_entry(update_request).await?;
-        Ok(result)
+    pub async fn _mm_delete_entry(&self, key_name: &str) -> Result<(), Error> {
+        let token = key_name.split("$").last().unwrap();
+        if token.eq("chunk") {
+            let response = self._delete_entry_chunk(key_name).await?;
+            return Ok(response);
+        }
+        return Err("invalid storage option".into());
     }
 }
